Moved: GameCommentServlet, ProfileCommentServlet, ProfileCreateServlet, ProfileDetailServlet, ProfileInfoEditServlet.
